Add Reservation repository, service, controller

+++ pool-service/src/main/java/test/ufanet/model/WorkingSchedule.java
@@ -15,7 +15,7 @@ import java.time.LocalTime;
 public class WorkingSchedule {
 
     @Id
-    private Short id;
+    private int id;
 
     @Column(name = "start_time", nullable = false)
     private LocalTime startTime;
